resolved circular dependencies of puzzle streak by using IList<PuzzleId> in PuzzleStreakResponse instead of Streak typedef, and extracting the notifier and provider into a new puzzle_streak_controller module
