+++ lib/src/model/puzzle/puzzle_repository.dart
@@ -14,7 +14,6 @@ import 'package:lichess_mobile/src/model/puzzle/puzzle.dart';
 import 'package:lichess_mobile/src/model/puzzle/puzzle_angle.dart';
 import 'package:lichess_mobile/src/model/puzzle/puzzle_difficulty.dart';
 import 'package:lichess_mobile/src/model/puzzle/puzzle_opening.dart';
-import 'package:lichess_mobile/src/model/puzzle/puzzle_streak.dart';
 import 'package:lichess_mobile/src/model/puzzle/puzzle_theme.dart';
 import 'package:lichess_mobile/src/model/puzzle/storm.dart';
 import 'package:lichess_mobile/src/network/http.dart';
@@ -222,7 +221,7 @@ sealed class PuzzleBatchResponse with _$PuzzleBatchResponse {
 sealed class PuzzleStreakResponse with _$PuzzleStreakResponse {
   const factory PuzzleStreakResponse({
     required Puzzle puzzle,
-    required Streak streak,
+    required IList<PuzzleId> streak,
 
     /// Timestamp of the response, used as streak unique identifier.
     ///
